pyproject.toml
[project]
name = "PARSER"
version = "0.1"
description = ""
authors = [
    {name = "maksyutov vlad", email = "maksyutov.vlad@gmail.com"},
]
dependencies = [
    "dependency-injector>=4.40.0",
    "SQLAlchemy>=1.4.44",
    "psycopg2>=2.9.5",
    "pydantic[email]>=2.0.2",
    "alembic>=1.8.1",
    "aiogram>=3.0.0b7",
    "setuptools>=65.6.3",
    "aiohttp>=3.8.3",
    "beautifulsoup4>=4.11.1",
    "scrapy>=2.7.1",
    "python-dotenv>=0.21.0",
    "psycopg2-binary>=2.9.5",
    "redis>=4.4.2",
    "sentry-sdk>=1.14.0",
    "sqlalchemy-utils>=0.40.0",
    "tesseract>=0.1.3",
    "pillow>=9.5.0",
    "pytesseract>=0.3.10",
    "langid>=1.1.6",
    "fastapi>=0.95.2",
    "strawberry-graphql[debug-server]>=0.178.0",
    "aioinject>=0.8.0",
    "business-validator>=2.0.0",
    "pyjwt>=2.7.0",
    "bcrypt>=4.0.1",
    "aio-pika>=9.2.2",
    "pydantic-settings>=2.0.3",
]
requires-python = ">=3.11"
license = {text = "MIT"}
[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
lint = { composite = [
    "ruff check . --fix --show-fixes",
    "black .",
    "ruff check . --fix --show-fixes",
    "mypy .",
] }


[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.2.0",
    "ruff>=0.0.192",
    "mypy>=0.991",
    "pytest-asyncio>=0.20.3",
    "anyio>=3.6.2",
    "black>=23.7.0",
]

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]
fixable = [
    "F",
    "E",
    "W",
    "I",
    "COM",
    "UP",
    "RUF",
]
ignore = [
    "E501", # Line Length
    "A003", # Class attribute ... is shadowing a python builtin
    "ANN101", # Missing type annotation for `self` in method
    "ANN102", # Missing type annotation for `cls` in classmethod
    "D", # pydocstyle
    "INP",
    "TCH001", # Move import into type-checking block
    "TCH003",
    "RUF012",
]
extend-exclude = [
    "*/__init__.py",
    "*/texts.py",
    "*/env.py"
]


[tool.pytest.ini_options]
pythonpath = "src"
asyncio_mode = "auto"
testpaths = ["tests"]
